+++ test/projection_test.dart
@@ -6,18 +6,20 @@ import 'package:winmd/winmd.dart';
 void main() {
   test('Simple int type', () {
     final type = TypeIdentifier(CorElementType.ELEMENT_TYPE_I4);
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.dartType(type), equals('int'));
-    expect(TypeBuilder.nativeType(type), equals('Int32'));
+    expect(typeProjection.dartType, equals('int'));
+    expect(typeProjection.nativeType, equals('Int32'));
   });
 
   test('Unicode string', () {
     final type = TypeIdentifier(CorElementType.ELEMENT_TYPE_PTR)
       ..name = 'LPWSTR'
       ..typeArgs = [TypeIdentifier(CorElementType.ELEMENT_TYPE_U2)];
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.dartType(type), equals('Pointer<Utf16>'));
-    expect(TypeBuilder.nativeType(type), equals('Pointer<Utf16>'));
+    expect(typeProjection.dartType, equals('Pointer<Utf16>'));
+    expect(typeProjection.nativeType, equals('Pointer<Utf16>'));
   });
 
   test('ANSI string', () {
@@ -25,9 +27,10 @@ void main() {
     final typedef = scope['Windows.Win32.KeyboardAndMouseInput.Apis']!;
     final api = typedef.findMethod('GetKeyNameTextA')!;
     final type = api.parameters[1].typeIdentifier; // LPSTR
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.dartType(type), equals('Pointer<Utf8>'));
-    expect(TypeBuilder.nativeType(type), equals('Pointer<Utf8>'));
+    expect(typeProjection.dartType, equals('Pointer<Utf8>'));
+    expect(typeProjection.nativeType, equals('Pointer<Utf8>'));
   });
 
   test('Handle', () {
@@ -35,9 +38,10 @@ void main() {
     final typedef = scope['Windows.Win32.KeyboardAndMouseInput.Apis']!;
     final api = typedef.findMethod('UnregisterHotKey')!;
     final type = api.parameters.first.typeIdentifier; // HWND
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.dartType(type), equals('int'));
-    expect(TypeBuilder.nativeType(type), equals('IntPtr'));
+    expect(typeProjection.dartType, equals('int'));
+    expect(typeProjection.nativeType, equals('IntPtr'));
   });
 
   test('Pointer<T>', () {
@@ -45,9 +49,10 @@ void main() {
     final typedef = scope['Windows.Win32.KeyboardAndMouseInput.Apis']!;
     final api = typedef.findMethod('GetKeyboardState')!;
     final type = api.parameters.first.typeIdentifier; // PBYTE
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.dartType(type), equals('Pointer<Uint8>'));
-    expect(TypeBuilder.nativeType(type), equals('Pointer<Uint8>'));
+    expect(typeProjection.dartType, equals('Pointer<Uint8>'));
+    expect(typeProjection.nativeType, equals('Pointer<Uint8>'));
   });
 
   test('Pointer<Pointer<T>>', () {
@@ -55,10 +60,10 @@ void main() {
     final typedef = scope['Windows.Win32.Security.Apis']!;
     final api = typedef.findMethod('CredReadW')!;
     final type = api.parameters.last.typeIdentifier; // PCREDENTIALW *
+    final typeProjection = TypeProjector(type);
 
-    expect(
-        TypeBuilder.nativeType(type), equals('Pointer<Pointer<CREDENTIAL>>'));
-    expect(TypeBuilder.dartType(type), equals('Pointer<Pointer<CREDENTIAL>>'));
+    expect(typeProjection.nativeType, equals('Pointer<Pointer<CREDENTIAL>>'));
+    expect(typeProjection.dartType, equals('Pointer<Pointer<CREDENTIAL>>'));
   });
 
   test('Unicode string w/ double pointer', () {
@@ -66,9 +71,10 @@ void main() {
     final typedef = scope['Windows.Win32.Shell.Apis']!;
     final api = typedef.findMethod('SHGetKnownFolderPath')!;
     final type = api.parameters.last.typeIdentifier; // PWSTR *
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.dartType(type), equals('Pointer<Pointer<Utf16>>'));
-    expect(TypeBuilder.nativeType(type), equals('Pointer<Pointer<Utf16>>'));
+    expect(typeProjection.dartType, equals('Pointer<Pointer<Utf16>>'));
+    expect(typeProjection.nativeType, equals('Pointer<Pointer<Utf16>>'));
   });
 
   test('Pass COM interfaces', () {
@@ -76,9 +82,10 @@ void main() {
     final typedef = scope['Windows.Win32.Com.Apis']!;
     final api = typedef.findMethod('CoSetProxyBlanket')!;
     final type = api.parameters.first.typeIdentifier; // IUnknown
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.nativeType(type), equals('Pointer'));
-    expect(TypeBuilder.dartType(type), equals('Pointer'));
+    expect(typeProjection.nativeType, equals('Pointer'));
+    expect(typeProjection.dartType, equals('Pointer'));
   });
 
   test('Pass pointers to COM interfaces', () {
@@ -86,9 +93,10 @@ void main() {
     final typedef = scope['Windows.Win32.Com.Apis']!;
     final api = typedef.findMethod('CoCreateInstance')!;
     final type = api.parameters[1].typeIdentifier; // LPUNKNOWN
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.nativeType(type), equals('Pointer'));
-    expect(TypeBuilder.dartType(type), equals('Pointer'));
+    expect(typeProjection.nativeType, equals('Pointer'));
+    expect(typeProjection.dartType, equals('Pointer'));
   });
 
   test('Pass double pointers to COM interfaces', () {
@@ -96,9 +104,10 @@ void main() {
     final typedef = scope['Windows.Win32.Automation.Apis']!;
     final api = typedef.findMethod('GetActiveObject')!;
     final type = api.parameters.last.typeIdentifier; // IUnknown **
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.nativeType(type), equals('Pointer<Pointer>'));
-    expect(TypeBuilder.dartType(type), equals('Pointer<Pointer>'));
+    expect(typeProjection.nativeType, equals('Pointer<Pointer>'));
+    expect(typeProjection.dartType, equals('Pointer<Pointer>'));
   });
 
   test('OLECHAR is represented correctly', () {
@@ -106,9 +115,10 @@ void main() {
     final typedef = scope['Windows.Win32.Automation.Apis']!;
     final api = typedef.findMethod('SysAllocString')!;
     final type = api.parameters.first.typeIdentifier; // OLECHAR *
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.nativeType(type), equals('Pointer<Utf16>'));
-    expect(TypeBuilder.dartType(type), equals('Pointer<Utf16>'));
+    expect(typeProjection.nativeType, equals('Pointer<Utf16>'));
+    expect(typeProjection.dartType, equals('Pointer<Utf16>'));
   });
 
   test('Callbacks are represented correctly', () {
@@ -116,10 +126,11 @@ void main() {
     final typedef = scope['Windows.Win32.Gdi.Apis']!;
     final api = typedef.findMethod('EnumFontFamiliesExW')!;
     final type = api.parameters[2].typeIdentifier; // FONTENUMPROCW
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.nativeType(type),
+    expect(typeProjection.nativeType,
         equals('Pointer<NativeFunction<EnumFontFamExProc>>'));
-    expect(TypeBuilder.dartType(type),
+    expect(typeProjection.dartType,
         equals('Pointer<NativeFunction<EnumFontFamExProc>>'));
   });
 
@@ -128,8 +139,9 @@ void main() {
     final typedef = scope['Windows.Win32.SystemServices.Apis']!;
     final api = typedef.findMethod('InitializeProcThreadAttributeList')!;
     final type = api.parameters.first.typeIdentifier; // FONTENUMPROCW
+    final typeProjection = TypeProjector(type);
 
-    expect(TypeBuilder.nativeType(type), equals('IntPtr'));
-    expect(TypeBuilder.dartType(type), equals('int'));
+    expect(typeProjection.nativeType, equals('IntPtr'));
+    expect(typeProjection.dartType, equals('int'));
   });
 }