+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.0-dev.8
+
+- Refactor code, change enums to return `Uint32` by default.
+
+## 0.4.0-dev.7
+
+- Rework API for projection of metadata
+
 ## 0.4.0-dev.6
 
 - Fix bug associated with vending winmd COM objects